ai-agents/requirements.txt
crewai>=0.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0 